Fix grouped views
Something changed in the emby returned paths so the verification was
failing. For now, we are getting one item from the media folder and
comparing using the user view to make sure we are referring to the
correct one with the tag name. Asked Luke for an api that would do this.