Resort to the expensive database lookup only if the person exists in the database.
The database is configured such that each person name within the table
has to be unique. Attempting to add a person which already exists causes
an IntegrityError exception to be thrown. This is a less costly
operation than the cursor.execute and cursor.fetch operations.

Increase performance by reverting to the expensive calls of fetching if
adding a person fails.

This pattern could also be applied to other tables within the database.