music: Fix multidisc indexing
Kodi uses internally a right-shift operation to codify in a
single integer disc number and track number. This is taken
into account in this addon but, incorrectly, it is done for
discnumber > 1 only, while Kodi uses it for tracks in all
discs.

Fix this by directly codifying the track and disc number
for all discs.

This patch is based on investigation by user @neightwulf in
github.